Refactoring of tests related with pmaps.
Tests related with pmaps have been refactored, introducing new tests
in the high level objects and removing some obsolete tests.
